Refactored entire script, now everything is manageable 😭.
It was becoming a brain drain jumping around everywhere
